revert: "NCN-109324: enforce NutanixMachine and NutanixFailureDomain spec field "subnets" not have duplicates (#560)"
Validation changes must not be introduced in a patch release.

This reverts commit b7a3a26ea418911c1eb2db039722ca335c462c6c.
