+++ config/crd/bases/infrastructure.cluster.x-k8s.io_nutanixfailuredomains.yaml
@@ -114,7 +114,6 @@ spec:
                       otherwise'
                     rule: 'self.type == ''uuid'' ? has(self.uuid) && self.uuid.contains(''-'')
                       : !has(self.uuid)'
-                maxItems: 32
                 minItems: 1
                 type: array
                 x-kubernetes-validations:
@@ -124,10 +123,6 @@ spec:
             - prismElementCluster
             - subnets
             type: object
-            x-kubernetes-validations:
-            - message: each subnet must be unique
-              rule: 'size(self.subnets) > 1 ? self.subnets.all(x, self.subnets.exists_one(y,
-                x == y)) : true'
           status:
             description: NutanixFailureDomainStatus defines the observed state of
               NutanixFailureDomain resource.
