+++ config/crd/bases/infrastructure.cluster.x-k8s.io_nutanixmachines.yaml
@@ -463,7 +463,6 @@ spec:
                       otherwise'
                     rule: 'self.type == ''uuid'' ? has(self.uuid) && self.uuid.contains(''-'')
                       : !has(self.uuid)'
-                maxItems: 32
                 type: array
               systemDiskSize:
                 anyOf:
@@ -494,9 +493,6 @@ spec:
             x-kubernetes-validations:
             - message: Either 'image' or 'imageLookup' must be set, but not both
               rule: has(self.image) != has(self.imageLookup)
-            - message: each subnet must be unique
-              rule: 'has(self.subnet) && size(self.subnet) > 1 ? self.subnet.all(x,
-                self.subnet.exists_one(y, x == y)) : true'
           status:
             description: NutanixMachineStatus defines the observed state of NutanixMachine
             properties:
