+++ config/crd/bases/infrastructure.cluster.x-k8s.io_nutanixmachinetemplates.yaml
@@ -490,7 +490,6 @@ spec:
                               and forbidden otherwise'
                             rule: 'self.type == ''uuid'' ? has(self.uuid) && self.uuid.contains(''-'')
                               : !has(self.uuid)'
-                        maxItems: 32
                         type: array
                       systemDiskSize:
                         anyOf:
@@ -523,9 +522,6 @@ spec:
                     - message: Either 'image' or 'imageLookup' must be set, but not
                         both
                       rule: has(self.image) != has(self.imageLookup)
-                    - message: each subnet must be unique
-                      rule: 'has(self.subnet) && size(self.subnet) > 1 ? self.subnet.all(x,
-                        self.subnet.exists_one(y, x == y)) : true'
                 required:
                 - spec
                 type: object
